+++ src/test/java/com/purej/vminspect/http/server/JettyServerTest.java
@@ -25,11 +25,11 @@ public final class JettyServerTest {
    */
   public static void main(String[] args) throws Exception {
     // Config from cmd or default:
-    int port = args.length > 0 ? Integer.parseInt(args[0]) : 8080;
-    int frequencyMs = args.length > 1 ? Integer.parseInt(args[1]) : 60000;
-    String storageDir = args.length > 2 ? args[2] : null;
+    var port = args.length > 0 ? Integer.parseInt(args[0]) : 8080;
+    var frequencyMs = args.length > 1 ? Integer.parseInt(args[1]) : 60000;
+    var storageDir = args.length > 2 ? args[2] : null;
 
-    Server server = new Server(port);
+    var server = new Server(port);
 
     // Init 1: Direct init
     // VmInspectionServlet servlet = new VmInspectionServlet();
@@ -38,15 +38,15 @@ public final class JettyServerTest {
     // ServletHolder servletHolder = new ServletHolder(servlet);
 
     // Init 2: servlet parameters
-    ServletHolder servletHolder = new ServletHolder(VmInspectionServlet.class);
+    var servletHolder = new ServletHolder(VmInspectionServlet.class);
+    servletHolder.setInitParameter("vminspect.mbeans.defaultDomainFilter", "purej.*");
     servletHolder.setInitParameter("vminspect.mbeans.readonly", "false");
     servletHolder.setInitParameter("vminspect.mbeans.writeConfirmation", "true");
-    // servletHolder.setInitParameter("vminspect.mbeans.accessControlFactory",
-    // CustomMBeanAccessControlFactory.class.getName());
+    // servletHolder.setInitParameter("vminspect.mbeans.accessControlFactory", CustomMBeanAccessControlFactory.class.getName());
     servletHolder.setInitParameter("vminspect.statistics.collection.frequencyMs", String.valueOf(frequencyMs));
     servletHolder.setInitParameter("vminspect.statistics.storage.dir", storageDir);
 
-    ServletContextHandler handler = new ServletContextHandler();
+    var handler = new ServletContextHandler();
     handler.setContextPath("/inspect");
     handler.addServlet(servletHolder, "/*");
     server.setHandler(handler);