+++ release-notes.md
@@ -1,5 +1,6 @@
 Version "current"
 - Upgraded to spring 3.3 and latest other dependencies
+- MBeanAccessControl allows setting a default domain filter
 
 Version 2.0.0 - (04.03.2023)
 - Upgraded to jakarta-servlet 6.0.0 which renames javax.servlet to jakarta.servlet packages